--- src/@types/Svg.ts
@@ -1,3 +1,0 @@
-export interface SvgProp {
-  className?: string;
-}
